pyproject.toml
[tool.poetry]
name = "django-telegram-bot-template"
version = "0.1.0"
description = ""
authors = ["Ilaletdinov Almaz <a.ilaletdinov@yandex.ru"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.9"
loguru = "^0.5.3"
pyTelegramBotAPI = "^4.2.0"
python-dotenv = "^0.19.2"
django-environ = "^0.8.1"
django-split-settings = "^1.1.0"
celery = "<5.0.0"
progressbar2 = "^3.55.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.1"
flake8-eradicate = "^1.1.0"
flake8-isort = "^4.0.0"
flake8-print = "^4.0.0"
flake8-docstrings = "^1.6.0"
pytest = "^6.2.2"
pytest-django = "^4.1.0"
python-dateutil = "^2.8.1"
requests-mock = "^1.9.3"
pytest-cov = "^3.0.0"
flake8-cognitive-complexity = "^0.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
